# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll/summary:
    get:
      operationId: getPayrollSummary
      tags: [payroll]
      summary: Preview a payroll calculation for a period, without creating a run
      description: >-
        The "Run Report" / draft-review step (PRD Section 6) — computes the same
        per-employee breakdown a `POST /payroll/runs` would store, without persisting
        anything. `period` is a shorthand (`this_week`/`last_week`/`this_month`,
        mirroring `artifacts/detail-hub/src/lib/payroll-data.ts`'s `getPeriodRange`);
        explicit `periodStart`+`periodEnd` win over `period` if both are given.
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [this_week, last_week, this_month]
        - name: periodStart
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: periodEnd
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: The computed payroll summary for the resolved period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSummaryResult"
        "400":
          description: Invalid period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payroll/runs:
    get:
      operationId: listPayrollRuns
      tags: [payroll]
      summary: List recent payroll runs
      description: Ordered most-recent-first. `limit` powers the Team hub's "3 most recent" view.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Recent payroll runs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRunResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPayrollRun
      tags: [payroll]
      summary: Run payroll for a period
      description: >-
        Computes and stores line items for the resolved period as a new `status:
        draft` run (PRD Section 6) — review-before-pay, nothing is actually paid out
        yet. See `PATCH /payroll/runs/{id}` to finalize.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollRunRequest"
      responses:
        "201":
          description: Draft payroll run created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRunResult"
        "400":
          description: Invalid request body or period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payroll/runs/{id}:
    get:
      operationId: getPayrollRun
      tags: [payroll]
      summary: Get a payroll run by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: The payroll run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRunResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No payroll run with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePayrollRun
      tags: [payroll]
      summary: Finalize a payroll run (mark it paid)
      description: >-
        `draft` -> `paid` only. This is a bookkeeping status flip, NOT a real payment
        execution — no direct deposit/check/ACH transfer actually happens (no payroll
        processor is connected anywhere in this codebase, PRD Section 9).
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePayrollRunRequest"
      responses:
        "200":
          description: Updated payroll run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRunResult"
        "400":
          description: Invalid request body, or the run isn't currently `draft`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No payroll run with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings, optionally within a date range and/or for one client
      description: >-
        Powers the calendar view (FR-10) — `start`/`end` (both `YYYY-MM-DD`, both
        optional, inclusive) filter on `date` rather than fetching every booking ever
        made. Omitting both returns every booking for the organization. `clientId`
        (Section 6.2 of PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md) further
        narrows the result to bookings for that client — it composes with (never
        replaces) the organization scope every request is already limited to, so a
        `clientId` belonging to a different organization returns an empty array
        rather than another organization's booking.
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      description: >-
        `packageIds` and `employeeSplit` fully describe the booking's line items and
        employee revenue split in one call (a `booking_packages`/`employee_splits`
        join row is written per entry). `employeeSplit` may be empty (a booking can be
        created before an employee is assigned), but if non-empty every percentage
        must be in `(0, 100]` and the set must sum to exactly 100.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: >-
            Invalid request body, an invalid `employeeSplit` (see this operation's
            description), or a `clientId`/`packageIds`/`employeeSplit[].employeeId`
            that doesn't reference a real row in this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/financials:
    get:
      operationId: getFinancialReport
      tags: [reports]
      summary: Full financial statement set for a period (Monthly/Quarterly/Annual Financial View)
      description: >-
        FR-10/FR-11/FR-13 of PRD_DetailHub_After_Package_Work.md Section 6.2. Computes
        an income statement (revenue -> COGS -> gross profit -> opex -> operating
        income -> interest/tax -> net income), a balance sheet, a cash flow statement,
        package mix, revenue-by-employee, and a trailing-6-calendar-month
        revenue/net-income trend (ending with the month containing `end`) for one
        `[start, end]` period. This is a deliberate hybrid: revenue, jobs, labor
        (real payroll gross for the period), package mix, revenue-by-employee, and
        accounts receivable are computed from real bookings/payroll/packages/
        employees data; fixed overhead, loan interest, depreciation, the illustrative
        25% tax rate, and every balance-sheet field without a real underlying ledger
        are ported verbatim as constants from the app's original mock financial model
        (`artifacts/detail-hub/src/lib/reporting-data.ts`) — see
        `calculateFinancialReport` in `@workspace/db` for the exact breakdown. Every
        ratio is guarded against a zero-revenue/zero-jobs period (FR-13): `0`, never
        `NaN`/`Infinity`.
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: The computed financial report for the requested period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialReportResult"
        "400":
          description: Invalid request (missing/malformed start or end, or start after end).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    PayrollSummaryResult:
      type: object
      description: >-
        Response of `GET /payroll/summary` — the same computation `POST /payroll/runs`
        stores, without persisting anything.
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PayrollLineItem"
        grossTotal:
          type: number
        netTotal:
          type: number
        pendingTimeOffCount:
          type: integer
      required:
        - periodStart
        - periodEnd
        - lines
        - grossTotal
        - netTotal
        - pendingTimeOffCount
    PayrollRunResult:
      type: object
      description: docs/prds/PRD_DetailHub_Payroll_Module.md Section 2.6.
      properties:
        id:
          type: integer
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        durationType:
          type: string
          enum: [weekly, biweekly, monthly, custom]
        status:
          type: string
          enum: [draft, processing, paid, failed]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PayrollLineItem"
        runBy:
          type: integer
          description: id of the employee record attributed as having run this payroll.
        runAt:
          type: string
          format: date-time
        reportUrl:
          type: string
          nullable: true
      required:
        - id
        - periodStart
        - periodEnd
        - durationType
        - status
        - lineItems
        - runBy
        - runAt
    CreatePayrollRunRequest:
      type: object
      description: >-
        Either `period` (shorthand) or explicit `periodStart`+`periodEnd` must resolve
        to a valid range — see `GET /payroll/summary`'s description for the same
        resolution rule.
      properties:
        period:
          type: string
          enum: [this_week, last_week, this_month]
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        durationType:
          type: string
          enum: [weekly, biweekly, monthly, custom]
        runByEmployeeId:
          type: integer
          description: >-
            Required — see `PayrollRunResult.runBy`'s description and
            `createPayrollRun`'s doc comment in `@workspace/db` for why this can't be
            inferred automatically from the authenticated session in v1.
      required:
        - durationType
        - runByEmployeeId
    UpdatePayrollRunRequest:
      type: object
      description: "Only `status: paid` is accepted, and only from a `draft` run."
      properties:
        status:
          type: string
          enum: [paid]
      required:
        - status
    BookingResult:
      type: object
      description: >-
        Field names/shape match `artifacts/detail-hub/src/lib/mock-data.ts`'s
        `Booking` interface (`date`, `startTime`, `address`, not the master PRD's
        `scheduledDate`/`scheduledStartTime`/`serviceAddress` — the mock interface is
        the source of truth per Section 7), plus `createdAt`/`updatedAt`/`createdBy`.
        `employeeIds` is derived from `employeeSplit` for convenience, not stored
        separately. (`gasMeterStatus`/`weatherSnapshot` were removed per FR-9 of
        PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md — dead placeholder
        columns never populated by a real integration; FR-11 says they come back for
        real later, see Fuel_Gauge_PRD.md.) `paymentMethod`/`paymentReference`/
        `paymentRecordedAt`/`refundStatus`/`refundReference` match
        `PRD_DetailHub_Payment_Methods.md` Section 7 exactly (superseding the earlier,
        rougher `paymentMethod`/`paymentNote` shape this API had before that PRD was
        reviewed — see `../../lib/db/src/schema/bookings.ts` for why) — settable only
        via `POST /bookings/{id}/payment` and `POST /bookings/{id}/refund`, never
        through `POST`/`PATCH /bookings`.
      properties:
        id:
          type: integer
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
        employeeIds:
          type: array
          items:
            type: integer
          description: Derived from `employeeSplit` (every `employeeId` present in the split).
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
        date:
          type: string
          format: date
        startTime:
          type: string
          description: "24-hour HH:MM, e.g. \"09:00\"."
        address:
          type: string
        depositAmount:
          type: number
        parkingCost:
          type: number
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
          enum: [zelle, venmo, cash, credit_card]
        paymentReference:
          type: string
          nullable: true
        paymentRecordedAt:
          type: string
          format: date-time
          nullable: true
        refundStatus:
          type: string
          nullable: true
          enum: [none, requested, completed]
        refundReference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: id of the admin/owner user who created this booking.
      required:
        - id
        - clientId
        - packageIds
        - employeeIds
        - employeeSplit
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
        - createdAt
        - updatedAt
        - createdBy
    CreateBookingRequest:
      type: object
      properties:
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
          default: []
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
          default: []
        date:
          type: string
          format: date
        startTime:
          type: string
        address:
          type: string
          minLength: 1
        depositAmount:
          type: number
          minimum: 0
        parkingCost:
          type: number
          minimum: 0
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
      required:
        - clientId
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
...
    FinancialReportResult:
      type: object
      description: >-
        Response of `GET /reports/financials`. See that path's own description, and
        `calculateFinancialReport` in `@workspace/db`'s `reports.ts`, for the exact
        real-data-vs-fixed-constant breakdown of every field below.
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        months:
          type: array
          description: Every "YYYY-MM" spanned by the requested period, inclusive.
          items:
            type: string
        revenue:
          type: number
          description: Real — completed bookings in the period x their assigned packages' current price.
        jobs:
          type: integer
          description: Real — count of completed bookings in the period.
        avgTicket:
          type: number
          description: "Real. Guarded against division by zero (FR-13): 0 when jobs is 0."
        cogs:
          type: object
          properties:
            labor:
              type: number
              description: Real — total gross payroll for the period (`calculatePayrollSummary`'s `grossTotal`).
            materials:
              type: number
              description: "Fixed rate (7.96%) applied to real revenue."
            gas:
              type: number
              description: "Fixed rate ($11.833/job) applied to real jobs."
            total:
              type: number
          required: [labor, materials, gas, total]
        grossProfit:
          type: number
        grossMarginPct:
          type: number
          description: "Guarded against division by zero (FR-13): 0 when revenue is 0."
        opex:
          type: object
          description: >-
            Fixed monthly overhead constants (insurance/vehicleMaint/software/
            marketing/adminWages), scaled by the number of calendar months `months`
            spans. No real overhead tracking exists yet.
          properties:
            insurance:
              type: number
            vehicleMaint:
              type: number
            software:
              type: number
            marketing:
              type: number
            adminWages:
              type: number
            total:
              type: number
          required: [insurance, vehicleMaint, software, marketing, adminWages, total]
        operatingIncome:
          type: number
        operatingMarginPct:
          type: number
          description: "Guarded against division by zero (FR-13): 0 when revenue is 0."
        interest:
          type: number
          description: Fixed ($140/mo, scaled by months spanned) — loan interest.
        preTaxIncome:
          type: number
        tax:
          type: number
          description: Fixed illustrative 25% rate applied to real (max(0, ...)) pre-tax income.
        netIncome:
          type: number
        netMarginPct:
          type: number
          description: "Guarded against division by zero (FR-13): 0 when revenue is 0."
        depreciation:
          type: object
          description: Fixed monthly depreciation constants, scaled by months spanned. Non-cash.
          properties:
            vehicle:
              type: number
            equipment:
              type: number
            total:
              type: number
          required: [vehicle, equipment, total]
        packageMix:
          type: array
          items:
            $ref: "#/components/schemas/PackageMixEntry"
        revenueByEmployee:
          type: array
          items:
            $ref: "#/components/schemas/RevenueByEmployeeEntry"
        balanceSheet:
          $ref: "#/components/schemas/FinancialReportBalanceSheet"
        cashFlow:
          $ref: "#/components/schemas/FinancialReportCashFlow"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/FinancialReportTrendPoint"
      required:
        - periodStart
        - periodEnd
        - months
        - revenue
        - jobs
        - avgTicket
        - cogs
        - grossProfit
        - grossMarginPct
        - opex
        - operatingIncome
        - operatingMarginPct
        - interest
        - preTaxIncome
        - tax
        - netIncome
        - netMarginPct
        - depreciation
        - packageMix
        - revenueByEmployee
        - balanceSheet
        - cashFlow
        - trend
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable detail, when available.
      required:
        - error